apps/api/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.30.0

# Validation & Config
pydantic==2.5.0
pydantic-settings==2.1.0

# Cache
redis==5.0.1

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==41.0.7

# HTTP & WebSocket
httpx==0.25.2
python-socketio==5.10.0
python-multipart==0.0.6

# Storage
boto3==1.29.7  # For MinIO (S3-compatible)

# Utilities
python-dotenv==1.0.0
email-validator==2.1.0

# Name Matching
fuzzywuzzy==0.18.0
python-Levenshtein==0.25.1
